custom/src/dst_data_builder/simple_requirements.txt
openai>=1.0.0
omegaconf>=2.3.0
hydra-core>=1.3.0
pytest>=7
